[LIB]: Copy correct length in sctp_getaddrs
When performing the memmove trick in sctp_getaddrs, do not adjust the
length.  This is needed because the kernel has always returned back
just the length of the packed address array, excluding the sctp_getaddrs
structure.

Thanks to Thomas Dreibholz <dreibh@iem.uni-due.de> for tracking down and
fixing this.

Signed-off-by: Vlad Yasevich <vladislav.yasevich@hp.com>