frame_test: provide a working test for fast retransmissions.
We've broken fast retransmits a few times recently and we really
needed a test for regression purposes just like that.  The existing
test was completely bogus in it's expected behavior and the way
it was written.  So this re-writes the test to make sure that
different fast retransmission scenarios are covered and that
we correctly implement fast recovery algorithm.

Signed-off-by: Vlad Yasevich <vladislav.yasevich@hp.com>