Move frame tests to a separate repository
Kernel developement has been acceleriting not just
in SCTP, but in other areas of the kernel.  Since
the frame tests depend on a large amount of kernel code,
they have been very fragile over the last year or so.
At times, they break due to a change in a completely
unrelated subsystem (like vm).  As such, I am moving
them to a separate repository so that they do not
hold back changes to SCTP library and user applications.

The new repository is available at:
    git://git.kernel.org/pub/scm/linux/kernel/git/vxy/lksctp-test.git

This new repository is aimed at kernel developers who wish
to extend the SCTP kernel implementation.

Signed-off-by: Vlad Yasevich <vladislav.yasevich@hp.com>