lib: Support non-blocking sctp_connectx() calls
1: I decided to keep the previous "new" sctp_connectx() implementation,
because a: it costs nothing, and b: some very recent programs may rely
on the broken behaviour.

2: The new-new sctp_connectx() tries the new method first. If it fails
due to ENOPROTOOPT (old kernel) then it falls back to the old method first, but
only if it is safe to do so (blocking socket, or assoc_id not returned)

3: (note of users of this feature): If a non-blocking connect fails
quickly then a new call to sctp_connectx() may return an identical association
id becaue the ID has been freed up already.

Additional slight code clean-up.

Signed-off-by: Ivan Skytte Jørgensen <isj-sctp@i1.dk>
Signed-off-by: Vlad Yasevich <vladislav.yasevich@hp.com>