api: Remove the sctp option enumerator
The sctp_optname enumerator is actually getting in the way as
we try to deprecate and remove really old options, because removing
items from the enum breaks binary compatibility with the kernel.
As a result, we do not gain any benefits from the enumerator, but
in fact get obfuscation of option values.  Now mirror the kernel
change and turn the enum into a defined list with specific values.

Signed-off-by: Vlad Yasevich <vladislav.yasevich@hp.com>