lksctp-tools: Add sctp_peeloff_flags function
With the addition of the SCTP_SOCKOPT_PEELOFF_FLAGS socket option for
linux, this patch adds a library function to support its use.  For non
linux systems, when the call is used with a 0 flag set, the behavior
will be identical to sctp_peeloff, and with a non-zero flag set for
non-linux operating systems, a runtime error will be returned.

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>
CC: Vlad Yasevich <vyasevic@redhat.com>
CC: Daniel Borkmann <dborkman@redhat.com>
Signed-off-by: Marcelo Ricardo Leitner <marcelo.leitner@gmail.com>