tests: fix SMALL_MAXSEG
After Commit ecca8f88da5c ("sctp: set frag_point in sctp_setsockopt_maxseg
correctly") in kernel, sctp will not allow to set max_seg with a too small
value.

Now it's using SMALL_MAXSEG in func_tests, with SMALL_MAXSEG = 100,
sctp_setsockopt always returns err and causes some tests to fail.

This patch is to avoid it by defining SMALL_MAXSEG as 500 instead, 500
is a good value, as it's very close to the minimum and will not cause
the last frag's size is different from the others', which would also
lead to failure of some tests.

Signed-off-by: Xin Long <lucien.xin@gmail.com>