sctp.h: make use kernel UAPI header
and with that, remove tons of duplicated declarations. These were
lagging behind the kernel header, which is always the most updated one,
and as the library is intended to be run only on Linux, there is no
reason that we cannot re-use what is in there.

That said, this patch drops all duplicated declarations and includes
linux/sctp.h to bring them into lksctp-tools.

Signed-off-by: Marcelo Ricardo Leitner <marcelo.leitner@gmail.com>