include: add private copy of kernel UAPI
Commit 3c8bd0d26b64 ("sctp.h: make use kernel UAPI header") removed the
duplicated bits between lksctp-tools exported sctp.h header and kernel
UAPI, but missed that the library should still build with kernels that
don't provide an updated UAPI one.

This commit thus adds a private copy of a recent kernel UAPI, which will
need to be synced every now and then, as before, but won't confuse
applications on which header to actually use.

Closes: #22
Closes: #23

Fixes: 3c8bd0d26b64 ("sctp.h: make use kernel UAPI header")
Signed-off-by: Marcelo Ricardo Leitner <marcelo.leitner@gmail.com>