build: remove internal copy of kernel header
The library should not use a static copy, as that creates problems on
how to conform to what the kernel expects. One example is the struct
sctp_event_subscribe, which the kernel only accepts buffers up to the size
it wants. Larger buffers, which would happen if an updated library is
built/ran under an old kernel, are rejected.

Subsequent patches will adjust the library so it can adapt itself to the
kernel-headers available at compile time.

Acked-by: Neil Horman <nhorman@tuxdriver.com>
Signed-off-by: Marcelo Ricardo Leitner <marcelo.leitner@gmail.com>