func_tests: free the allocated buffers before return
All resource allocated should be freed before return from a function,
even in the main function.

This patch only frees the allocated buffers when a test succeeds, and
not take care of the test failure where it calls DUMP_CORE to exit.

While at it, also fix some indents around.

Signed-off-by: Xin Long <lucien.xin@gmail.com>